🚚 (phi-das): Rename service dir to services
