+++ apps/phi-das/src/routers/clinicalProfiles.ts
@@ -19,8 +19,8 @@
 
 import { Router } from 'express';
 
-import { getClinicalProfile, getClinicalProfiles } from '../service/search.js';
-import { createClinicalProfile } from '../service/create.js';
+import { getClinicalProfile, getClinicalProfiles } from '../services/search.js';
+import { createClinicalProfile } from '../services/create.js';
 import logger from '../logger.js';
 
 // TODO: update JSDoc comments when custom error handling is implemented
